+++ x/posts/internal/types/genesis.go
@@ -1,17 +1,21 @@
 package types
 
+import "fmt"
+
 // GenesisState contains the data of the genesis state for the posts module
 type GenesisState struct {
-	Posts       Posts                    `json:"posts"`
-	PollAnswers map[string]UserAnswers   `json:"poll_answers_details"`
-	Reactions   map[string]PostReactions `json:"reactions"`
+	Posts       Posts                  `json:"posts"`
+	PollAnswers map[string]UserAnswers `json:"poll_answers_details"`
+	Reactions   map[string]Reactions   `json:"reactions"`
+	Hashtags    map[string]PostIDs     `json:"hashtags"`
 }
 
 // NewGenesisState creates a new genesis state
-func NewGenesisState(posts Posts, reactions map[string]PostReactions) GenesisState {
+func NewGenesisState(posts Posts, reactions map[string]Reactions, hashtags map[string]PostIDs) GenesisState {
 	return GenesisState{
 		Posts:     posts,
 		Reactions: reactions,
+		Hashtags:  hashtags,
 	}
 }
 
@@ -40,6 +44,14 @@ func ValidateGenesis(data GenesisState) error {
 		for _, record := range reactions {
 			if err := record.Validate(); err != nil {
 				return err
+			}
+		}
+	}
+
+	for _, IDs := range data.Hashtags {
+		for _, id := range IDs {
+			if !id.Valid() {
+				return fmt.Errorf("invalid post ID, %s", id)
 			}
 		}
 	}